+++ client/apps.go
@@ -1,6 +1,7 @@
 package client
 
 import (
+	"fmt"
 	"io"
 	"os"
 	"strconv"
@@ -62,4 +63,11 @@ func (c *Client) GetEnvironment(name string) (pb.Environment, error) {
 func (c *Client) SetEnvironment(name string, data string) error {
 	_, err := c.ProviderServiceClient.EnvironmentSet(ctx, &pbs.EnvSetRequest{Name: name, Data: data})
 	return err
+}
+
+func (c *Client) ProxyRemote(host string, port int, conn io.ReadWriteCloser) error {
+	return c.Stream("/ws/proxy", map[string]string{
+		"host": host,
+		"port": fmt.Sprintf("%d", port),
+	}, conn, conn)
 }